finalize unit tests
